implement fetching/creating accounts and Transactions

+++ JobRunner/SmartMoneyJobRunner/SmartMoneyJobRunner/Models/Stop.cs
@@ -8,7 +8,7 @@ namespace SmartMoneyJobRunner.Models
         public DbGeography Location { get; set; }
         public Guid UserId { get; set; }
         public Guid Id { get; set; }
-        public int Duration { get; set; }
+        public decimal Duration { get; set; }
         public Stop()
         {
             Id = Guid.NewGuid();
